requirements.txt
streamlit==1.39.0
numpy==1.26.4
scipy==1.14.1
sounddevice==0.5.0
gTTS==2.5.3
groq==0.11.0
python-dotenv==1.0.1